+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8">
-    <link rel="icon" href="/favicon.ico">
+    <link rel="icon" href="/favicon.ico"> <!-- Ensure path starts with / -->
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>YourBank</title>
     <!-- Remove Tailwind CDN -->
