+++ src/pages/GetStartedPage.vue
